% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCues.R
\name{CODE}
\alias{CODE}
\title{Helper function for makeCues}
\usage{
CODE(SPEC, num)
}
\arguments{
\item{SPEC}{Spectrum representation made in makeCues()}

\item{num}{Number of the part}
}
\value{
A string containing the coding. Each band is seperated by "_".
}
\description{
Helper function for makeCues
}
\author{
Denis Arnold
}
