% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotation.R
\name{readTextGridRobust}
\alias{readTextGridRobust}
\title{Reads a TextGrid made with praat and returns a list with a vector of all tier names and a data.frame for each tier}
\usage{
readTextGridRobust(File, Encoding)
}
\arguments{
\item{File}{Name (with full path, if not in wd) of the TextGrid}

\item{Encoding}{Encoding of the TextGrid. Typically encodings are "ACSII","UTF-8" or "UTF-16"}
}
\value{
A list containing a vectors with the names and data.frames for each tier in the TextGrid.
}
\description{
Reads a TextGrid made with praat and returns a list with a vector of all tier names and a data.frame for each tier
}
\examples{
       \dontrun{
       # Assume that NameOfTextGrid is encoded in "UTF-8"
       Data=readTextGridRobust("NameOfTextGrid","UTF-8")

       }
}
\author{
Denis Arnold
}
