% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotation.R
\name{readWavesurfer}
\alias{readWavesurfer}
\title{Reads a New Wavesurfer style annotation file and returns a data.frame with times and lables}
\usage{
readWavesurfer(File, Encoding)
}
\arguments{
\item{File}{Name (with full path, if not in wd) of the annotation file}

\item{Encoding}{Encoding of the annotation file. Typically encodings are "ACSII","UTF-8" or "UTF-16"}
}
\value{
A data.frame with $Output for the lable $start and $end time of the lable.
}
\description{
Reads a New Wavesurfer style annotation file and returns a data.frame with times and lables
}
\examples{
       \dontrun{
       # Assume that NameOfAnnotation is encoded in "UTF-8"
       Data=readWavesurfer("NameOfTextGrid","UTF-8")
       }
}
\author{
Denis Arnold
}
